Add new account working. Validations on its way

+++ src/com/bankapp/dao/UserDAO.java
@@ -16,7 +16,7 @@ public interface UserDAO {
 	public void insert(UserInfo user);
 	public List<UserInfo> getUserList();
 	public UserInfo findUserByUsername(String username);
-	public int registerNewUserAccount(UserInfo userInfo, Useraccounts account);
+	public Long registerNewUserAccount(UserInfo userInfo, Useraccounts account);
 	public String findUserRoleType(String username);
 	public void updateUserInfo(UserInfo userInfo);
 	public void deleteUserInfo(UserInfo userInfo);
